Hack-fix the transaction time for BTC
Outbound transactions are received by the blockchain at received_at
and the block is confirmed at a later time.

If we use the block confirmation time as the time of the outbound
transaction, there is a chance that the inbound transaction will
have already been registered - unsure why the inbound transaction
does not occur after the block has been validated, I probably do
not interpret these timestamps well.

In any case, that's problematic to register the outbound
transactions with a delay, because it may break the assumption
that an outbound transaction has a timestamp prior to the inbound
one.

For that reason, we move outbound transactions timestamps an hour
earlier.